--- oauth-2.0/authzserver/src/main/java/org/apache/oltu/oauth2/issuer/OAuthIssuer.java
@@ -1,37 +1,0 @@
-/**
- *       Copyright 2010 Newcastle University
- *
- *          http://research.ncl.ac.uk/smart/
- *
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.oltu.oauth2.issuer;
-
-import org.apache.oltu.oauth2.common.exception.OAuthSystemException;
-
-/**
- *
- *
- *
- */
-public interface OAuthIssuer {
-    public String accessToken() throws OAuthSystemException;
-
-    public String authorizationCode() throws OAuthSystemException;
-
-    public String refreshToken() throws OAuthSystemException;
-}
