+++ oauth-2.0/authzserver/src/test/java/org/apache/oltu/oauth2/as/UUIDValueGeneratorTest.java
@@ -21,7 +21,7 @@
 
 package org.apache.oltu.oauth2.as;
 
-import org.apache.oltu.oauth2.issuer.UUIDValueGenerator;
+import org.apache.oltu.oauth2.as.issuer.UUIDValueGenerator;
 import org.junit.Assert;
 import org.junit.Test;
